feat: Loss-based sampling (sample-then-batch only) (#4)
This adds a first implementation of sample-then-batch loss-based
sampling. Critical follow up tasks are:

- Clean up the interface of the SelectiveSampler. There are many hooks
which might not be needed, and some even conflict ("after batch" without
the proper actually needed information).
- Generalize the full_finetune recipe to support both sample-then-batch
OR batch-then-sample using the same selective sampler. Right now we
hardcode sample-then-batch.


I also touch on the `set_num_selected_samples` stuff because it kept
giving me trouble and I did not see why it was necessary to do it like
that instead of based on the mask.

---------

Co-authored-by: Titus Robroek <titr@itu.dk>